fix: Use size_t instead of int for DownloadProgressIndicator constructor step count

+++ source/cli/ui/downloadprogressindicator.d
@@ -18,7 +18,7 @@ class DownloadProgressIndicator : Status {
         string _prevReport;
     }
 
-    this(in int stepCount) {
+    this(in size_t stepCount) {
         auto width = getTerminalWidth();
         _stepCount = stepCount;
